pyproject.toml
[project]
name = "typecast-api-mcp-server"
version = "0.1.0"
description = "MCP Server for typecast-api"
authors = [{ name = "Hyunseung Lee", email = "hyunseung@neosapience.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp",
    "httpx",
    "sounddevice",
    "soundfile",
    "click",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
typecast-api-mcp-server = "app.main:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = ["mcp[cli]>=1.6.0"]

[tool.hatch.build]
include = ["app/**/*.py", "README.md", "LICENSE"]
